+++ src/model/food.rs
@@ -1,9 +1,9 @@
 pub struct Food {
-    pub calories: u32,
+    pub calories: u16,
 }
 
 impl Food {
-    pub fn new(calories: u32) -> Self {
+    pub fn new(calories: u16) -> Self {
         Self { calories }
     }
 }
